상품등록 페이지 컴포넌트로 분리
